Add files via upload
Same version as before essentially but I deleted a lot of comments. I changed the "Lebron" place holder variable name to Diff_N2N, for difference between N and 2N. I also got rid of the 3 different ways to calculate max error and just kept the one line version of it.

I need to just move on to the second question and making plots and tables.