Add files via upload
This is the same as what I just uploaded but I changed a print statement and rounded the Log2(Error_N/ and Error 2N) equations.

The comment from before was:

Added part 2, Neuman BCs. I went ahead and made a new funtion for the Helmotz Thomas Algorithm function, rather than modify the first one with condition statements. I did this so that the function closely follows the pseudo code. I also changed the interval function back to how I originally had it because I need to approximate the u(x=0) value for the Neumann problem. I also added some print statements and enabled the outter for loop for the running the different k values.

Now I just gotta do the plotting and tables. I am not sure if I will add this stuff to the code with conditional statements or if I will just use my functions and make the plots in the console.